% Generated by roxygen2 (4.0.2): do not edit by hand
\name{otherViaPipe}
\alias{otherViaPipe}
\title{Example of using pipe on package-supplied script}
\usage{
otherViaPipe()
}
\value{
A data.frame read from the output
}
\description{
Example of calling a script via pip
}
\details{
Uses \code{system.file()} to portably obtain the path
of a shell script and uses \code{pipe()} to execute it,
using the command output as input to read.
}
\author{
Dirk Eddelbuettel
}

