% Generated by roxygen2 (4.0.2): do not edit by hand
\name{otherViaSystem}
\alias{otherViaSystem}
\title{Example of using system on package-supplied script}
\usage{
otherViaSystem()
}
\value{
Several lines of text
}
\description{
Example of calling a script via system
}
\details{
Uses \code{system.file()} to portably obtain the path
of a shell script and uses \code{system()} to execute.
}
\author{
Dirk Eddelbuettel
}

